requirements.txt
customtkinter==5.2.2
Pillow==10.4.0
python-dotenv==1.0.1
Requests==2.32.3
websockets==12.0
